requirements.txt
beautifulsoup4==4.6.3
google-cloud-firestore==0.30.0
lxml==4.4.2
requests==2.20.1